+++ src/utils/time.test.ts
@@ -25,7 +25,8 @@ describe("getValidTime", () => {
 
   it("load - numbers", () => {
     const now = new Date().getTime();
-    const dates = new Array(10000).fill(0).map((d) => {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const dates = new Array(10000).fill(0).map((_d) => {
       return Math.floor(Math.random() * now);
     });
 
@@ -33,14 +34,14 @@ describe("getValidTime", () => {
     dates.forEach((d) => getValidTime(d, "utc"));
     const end = new Date().valueOf();
     const operationLength = end - start;
-    console.log(`Validate dates: ${operationLength} ms`);
-
+    console.warn(`Validate dates: ${operationLength} ms`);
     expect(operationLength).toBeLessThan(500);
   });
 
   it("load - strings", () => {
     const now = new Date().getTime();
-    const dates = new Array(10000).fill(0).map((d) => {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const dates = new Array(10000).fill(0).map((_d) => {
       const millis = Math.floor(Math.random() * now);
       return new Date(millis).toISOString();
     });
@@ -49,8 +50,7 @@ describe("getValidTime", () => {
     dates.forEach((d) => getValidTime(d, "utc"));
     const end = new Date().valueOf();
     const operationLength = end - start;
-    console.log(`Validate dates: ${operationLength} ms`);
-
+    console.warn(`Validate dates: ${operationLength} ms`);
     expect(operationLength).toBeLessThan(500);
   });
 });